fix: library search boosting
Closes #1106
